Added the screen "new Flight"
I added a screen where the user can input two dates (startDate and endDate) as well as the origin state and departure state. The output will be all the flights that meet those requirements. The user will be able to scroll up and down to see all the flights. They will also be able to sort by date, carrier, origin airport or destination airport.
I also added an inputFormat parameter to the class inputBox (it is drawn on top of the input box) in order to inform about the expected format and information(s) from the input.